# SwaggerOffLine.yaml
openapi: 3.1.0
info:
  title: SpecRadar
  description: "API de inteligência competitiva automotiva - padroniza fichas técnicas\
    \ de veículos a partir de atributos livres, usando um catálogo canônico e um serviç\
    o de Machine Learning para mapear os atributos solicitados."
  contact:
    name: FIAP - Desafio Ford 01
    email: specradar@email.com.br
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearer-key: []
...
paths:
...
  /atributos/{id}:
    get:
      tags:
      - Atributos
      summary: Detalha um atributo pelo ID
      operationId: detalharAtributo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAtributo"
    put:
      tags:
      - Atributos
      summary: Atualiza um atributo do catálogo
      operationId: atualizarAtributo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAtualizacaoAtributo"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAtributo"
    delete:
      tags:
      - Atributos
      summary: Exclui (logicamente) um atributo do catálogo
      operationId: excluirAtributo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /usuarios:
    get:
      tags:
      - Usuários
      summary: Lista os usuários ativos de forma paginada
      operationId: listarUsuarios
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "nome,ASC"
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageDadosListagemUsuario"
    post:
      tags:
      - Usuários
      summary: Cadastra um novo usuário
      operationId: cadastrarUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroUsuario"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoUsuario"
  /pesquisas:
    get:
      tags:
      - Pesquisas
      summary: Lista as pesquisas realizadas pelo usuário autenticado
      operationId: listarPesquisas
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "criadaEm,ASC"
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageDadosListagemPesquisa"
    post:
      tags:
      - Pesquisas
      summary: Executa uma pesquisa e retorna a ficha técnica padronizada do veículo
      operationId: pesquisar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroPesquisa"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FichaTecnica"
  /login:
    post:
      tags:
      - Login
      summary: Autentica um usuário e retorna um token JWT válido por 60 minutos
      operationId: efetuarLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosLogin"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosTokenJWT"
      security: []
  /atributos:
    get:
      tags:
      - Atributos
      summary: "Lista os atributos ativos do catálogo, de forma paginada"
      operationId: listarAtributos
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "nome,ASC"
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageDadosDetalhamentoAtributo"
    post:
      tags:
      - Atributos
      summary: Cadastra um novo atributo no catálogo canônico
      operationId: cadastrarAtributo
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroAtributo"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAtributo"
  /usuarios/{id}/perfil:
    patch:
      tags:
      - Usuários
      summary: Altera o perfil (ADMIN ou ANALISTA) de um usuário
      operationId: atualizarPerfil
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAtualizacaoPerfil"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoUsuario"
  /usuarios/{id}:
    get:
      tags:
      - Usuários
      summary: Detalha um usuário pelo ID
      operationId: detalharUsuario
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoUsuario"
    delete:
      tags:
      - Usuários
      summary: Exclui (logicamente) um usuário
      operationId: excluirUsuario
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /pesquisas/{id}:
    get:
      tags:
      - Pesquisas
      summary: Detalha uma pesquisa do usuário autenticado (404 se pertencer a outro
        usuário)
      operationId: detalharPesquisa
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FichaTecnica"
components:
  schemas:
...
    DadosAtualizacaoAtributo:
      type: object
      properties:
        nome:
          type: string
          maxLength: 150
          minLength: 0
        unidade:
          type: string
          maxLength: 30
          minLength: 0
        categoria:
          type: string
          maxLength: 60
          minLength: 0
    DadosDetalhamentoAtributo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        codigo:
          type: string
        nome:
          type: string
        unidade:
          type: string
        categoria:
          type: string
        ativo:
          type: boolean
...
    DadosCadastroUsuario:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        login:
          type: string
          maxLength: 150
          minLength: 0
        senha:
          type: string
          maxLength: 100
          minLength: 8
        perfil:
          type: string
          enum:
          - ADMIN
          - ANALISTA
      required:
      - login
      - nome
      - perfil
      - senha
    DadosDetalhamentoUsuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        login:
          type: string
        perfil:
          type: string
          enum:
          - ADMIN
          - ANALISTA
        ativo:
          type: boolean
    DadosCadastroPesquisa:
      type: object
      properties:
        marca:
          type: string
          maxLength: 80
          minLength: 0
        modelo:
          type: string
          maxLength: 80
          minLength: 0
        versao:
          type: string
          maxLength: 80
          minLength: 0
        atributos:
          type: array
          items:
            type: string
            maxLength: 60
            minLength: 0
          maxItems: 50
          minItems: 1
      required:
      - atributos
      - marca
      - modelo
      - versao
    EspecificacaoResultado:
      type: object
      properties:
        atributoSolicitado:
          type: string
        atributoCanonico:
          type: string
        valor:
          type: string
        unidade:
          type: string
        status:
          type: string
          enum:
          - ENCONTRADO
          - NAO_DISPONIVEL
        confianca:
          type: number
          format: double
    FichaTecnica:
      type: object
      properties:
        veiculo:
          $ref: "#/components/schemas/VeiculoResumo"
        especificacoes:
          type: array
          items:
            $ref: "#/components/schemas/EspecificacaoResultado"
        geradoEm:
          type: string
          format: date-time
    VeiculoResumo:
      type: object
      properties:
        marca:
          type: string
        modelo:
          type: string
        versao:
          type: string
    DadosLogin:
      type: object
      properties:
        login:
          type: string
          minLength: 1
        senha:
          type: string
          minLength: 1
      required:
      - login
      - senha
    DadosTokenJWT:
      type: object
      properties:
        tokenJWT:
          type: string
    DadosCadastroAtributo:
      type: object
      properties:
        codigo:
          type: string
          maxLength: 60
          minLength: 0
        nome:
          type: string
          maxLength: 150
          minLength: 0
        unidade:
          type: string
          maxLength: 30
          minLength: 0
        categoria:
          type: string
          maxLength: 60
          minLength: 0
      required:
      - categoria
      - codigo
      - nome
    DadosAtualizacaoPerfil:
      type: object
      properties:
        perfil:
          type: string
          enum:
          - ADMIN
          - ANALISTA
      required:
      - perfil
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    DadosListagemUsuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        login:
          type: string
        perfil:
          type: string
          enum:
          - ADMIN
          - ANALISTA
    PageDadosListagemUsuario:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosListagemUsuario"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    DadosListagemPesquisa:
      type: object
      properties:
        id:
          type: integer
          format: int64
        marca:
          type: string
        modelo:
          type: string
        versao:
          type: string
        criadaEm:
          type: string
          format: date-time
    PageDadosListagemPesquisa:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosListagemPesquisa"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageDadosDetalhamentoAtributo:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosDetalhamentoAtributo"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT